deal with addresses that haven't been cluster by walletexplorer.com
There’s a small number of addresses not clustered by the API. We can
just assign a placeholder label to those. OTOH, it we see many of these
in a row, that is concerning, so we’ll die in this case. This commit
passes all tests.

No unit tests written for the changes as of yet :-/ (TODO)

Also, my text editor decided to fix some whitespace issues
automatically.